refactor type def
